NHSO-6305: App Camp Backend Refactor
Work done at July 2019 App Camp to refactor the backend:

- Ensure all projects are performing FxCop code analysis and have TreatWarningsAsErrors turned on.
- Suppressing CA1056 on SJR Appointments InformaticaUri as discussed with Cell C.
- Add AssemblyInfo.cs to SJR Models project.
- Update NDepend Project to analyse all non-unit-test projects.
- Update sln file to Visual Studio 16 (2019) format.
- Moved VersionTag.cs to auditing folder.
- Moved Audit functionality to its own assembly NHSOnline.Backend.Auditing.
- Moved auditor unit tests outside of support unit tests project to own project NHSOnline.Backend.Auditing.UnitTests.
- Moved AuditorTitles to AuditingOperations.
- Fixed namespace of ApiResponse
- Moved Im1CacheService from Support to GpSystems project.
- Moved Mongo session cache service to PfsApi
- Consolidated ApiSupport into Support.
- Moved OdsCodeLookup to GpSystems.

+++ backendworker/NHSOnline.Backend.GpSystems.UnitTests/Suppliers/Emis/Linkage/EmisLinkageServiceTests.cs
@@ -7,7 +7,6 @@ using AutoFixture.AutoMoq;
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
-using NHSOnline.Backend.CidApi.Areas.Linkage;
 using NHSOnline.Backend.GpSystems.Im1Connection;
 using NHSOnline.Backend.GpSystems.Linkage.Models;
 using NHSOnline.Backend.GpSystems.Linkage;
@@ -16,7 +15,6 @@ using NHSOnline.Backend.GpSystems.Suppliers.Emis.Linkage;
 using NHSOnline.Backend.GpSystems.Suppliers.Emis.Models;
 using NHSOnline.Backend.GpSystems.Suppliers.Emis.Models.Verifications;
 using NHSOnline.Backend.GpSystems.Suppliers.Emis.Session;
-using NHSOnline.Backend.Support;
 
 namespace NHSOnline.Backend.GpSystems.UnitTests.Suppliers.Emis.Linkage
 {
@@ -399,12 +397,11 @@ namespace NHSOnline.Backend.GpSystems.UnitTests.Suppliers.Emis.Linkage
 
             var mockResponse = new EmisClient.EmisApiObjectResponse<AddNhsUserResponse>(httpStatusCodeResponse);
 
-            // Add emis specific error code if unit test requires it.
+            // Add EMIS specific error code if unit test requires it.
             mockResponse.StandardErrorResponse = new StandardErrorResponse
             {
-                InternalResponseCode = (int)emisApiErrorCode,
+                InternalResponseCode = emisApiErrorCode,
             };
-
 
             _emisClient.Setup(x => x.NhsUserPost(
                 It.Is<EmisHeaderParameters>(
